# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jazment Cricket Prediction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get current user wallet
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/deposit:
    post:
      operationId: depositWallet
      tags: [wallet]
      summary: Deposit to wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletTransactionRequest"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/withdraw:
    post:
      operationId: withdrawWallet
      tags: [wallet]
      summary: Withdraw from wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletTransactionRequest"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

  # ─── Matches ──────────────────────────────────────────────────────────────────
...
  /withdrawals:
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create a withdrawal request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWithdrawalRequest"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/my:
    get:
      operationId: getMyWithdrawals
      tags: [withdrawals]
      summary: Get current user withdrawal history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalListResponse"
  /admin/withdrawals:
    get:
      operationId: adminListWithdrawals
      tags: [withdrawals, admin]
      summary: List all withdrawals (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Withdrawal list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminWithdrawalListResponse"
  /admin/withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveWithdrawal
      tags: [withdrawals, admin]
      summary: Approve a withdrawal and deduct wallet (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRejectRequest"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [withdrawals, admin]
      summary: Reject a withdrawal (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRejectRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Notifications ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Wallet:
      type: object
      properties:
        userId:
          type: string
        balance:
          type: number
        depositTotal:
          type: number
        withdrawTotal:
          type: number
        winTotal:
          type: number
        bonusTotal:
          type: number
      required:
        - userId
        - balance
        - depositTotal
        - withdrawTotal
        - winTotal
        - bonusTotal
    WalletTransactionRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
        note:
          type: string
      required:
        - amount
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [deposit, withdraw, win, loss, bonus, refund]
        amount:
          type: number
        balanceBefore:
          type: number
        balanceAfter:
          type: number
        status:
          type: string
          enum: [completed, pending, reversed]
        referenceId:
          type: string
        note:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - amount
        - balanceAfter
        - createdAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transactions
        - total
        - page
        - limit

    # ─── Match ─────────────────────────────────────────────────────────────────
...
    BankAccount:
      type: object
      properties:
        accountNumber:
          type: string
        ifsc:
          type: string
        holderName:
          type: string
    CreateWithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 500
        upiId:
          type: string
        bankAccount:
          $ref: "#/components/schemas/BankAccount"
      required:
        - amount
    Withdrawal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        amount:
          type: number
        upiId:
          type: string
        bankAccount:
          $ref: "#/components/schemas/BankAccount"
        status:
          type: string
          enum: [pending, approved, rejected]
        remarks:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - status
        - createdAt
        - updatedAt
    WithdrawalListResponse:
      type: object
      properties:
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - withdrawals
        - total
        - page
        - limit
    AdminWithdrawalUser:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
        walletBalance:
          type: number
    AdminWithdrawal:
      allOf:
        - $ref: "#/components/schemas/Withdrawal"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/AdminWithdrawalUser"
    AdminWithdrawalListResponse:
      type: object
      properties:
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/AdminWithdrawal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - withdrawals
        - total
        - page
        - limit

    # ─── Notifications ─────────────────────────────────────────────────────────
...
    ApproveRejectRequest:
      type: object
      properties:
        remarks:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        balanceAfter:
          type: number
      required:
        - success